rename: コンポーネント名を修正
